[add] Speech Synthesis model [optimize] upgrade to Node.js 20, PNPM 9, Husky 9 & other latest Upstream packages [fix] Example code bugs in Read Me document
