+++ .husky/pre-commit
@@ -1,5 +1,1 @@
-#!/bin/sh
-
-. "$(dirname "$0")/_/husky.sh"
-
 npm test
