+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "mobx-i18n",
-    "version": "0.5.0",
+    "version": "0.6.0",
     "license": "LGPL-3.0",
     "author": "shiy2008@gmail.com",
     "description": "Responsive Translation utility based on TypeScript & MobX",
@@ -27,25 +27,26 @@
     "module": "dist/index.esm.js",
     "main": "dist/index.js",
     "dependencies": {
-        "@swc/helpers": "^0.5.3",
-        "regenerator-runtime": "^0.14.1"
+        "@swc/helpers": "^0.5.13",
+        "regenerator-runtime": "^0.14.1",
+        "web-utility": "^4.4.0"
     },
     "peerDependencies": {
         "mobx": ">=6.11"
     },
     "devDependencies": {
-        "@parcel/config-default": "~2.11.0",
-        "@parcel/packager-ts": "~2.11.0",
-        "@parcel/transformer-typescript-tsc": "~2.11.0",
-        "@parcel/transformer-typescript-types": "~2.11.0",
-        "husky": "^8.0.3",
-        "lint-staged": "^15.2.0",
-        "mobx": "^6.12.0",
-        "parcel": "~2.11.0",
-        "prettier": "^3.2.4",
-        "typedoc": "^0.25.7",
-        "typedoc-plugin-mdn-links": "^3.1.12",
-        "typescript": "~5.3.3"
+        "@parcel/config-default": "~2.12.0",
+        "@parcel/packager-ts": "~2.12.0",
+        "@parcel/transformer-typescript-tsc": "~2.12.0",
+        "@parcel/transformer-typescript-types": "~2.12.0",
+        "husky": "^9.1.6",
+        "lint-staged": "^15.2.10",
+        "mobx": "^6.13.3",
+        "parcel": "~2.12.0",
+        "prettier": "^3.3.3",
+        "typedoc": "^0.26.8",
+        "typedoc-plugin-mdn-links": "^3.3.2",
+        "typescript": "~5.6.2"
     },
     "prettier": {
         "singleQuote": true,
@@ -63,7 +64,7 @@
         }
     },
     "scripts": {
-        "prepare": "husky install",
+        "prepare": "husky",
         "test": "lint-staged",
         "build": "rm -rf dist/ docs/  &&  parcel build  &&  typedoc source/",
         "prepublishOnly": "npm test  &&  npm run build"
